requirements.txt
# This list of dependencies was generated with `pip3 freeze`

# Logging and parse args
absl-py==1.4.0

# Gym environment
gym==0.25.2
ale-py==0.7.5
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
mujoco==2.2.2
box2d-py==2.3.5
imageio==2.31.1
pyglet==2.0.9
pygame==2.1.0

# PyTorch
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2

# Tensorboard
tensorboard==2.13.0

# Utilities
numpy==1.24.3
opencv-python==4.8.0.74
python-snappy==0.6.1

# Coding styles and others
pre-commit==3.3.2
black==23.7.0